requirements.txt
streamlit
easyocr
pillow
snowflake-connector-python
pandas
python-dotenv
